Refactoring harness integeration test files
