+++ test/src/main/java/org/apache/accumulo/test/harness/AccumuloClusterConfiguration.java
@@ -16,10 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.accumulo.harness.conf;
+package org.apache.accumulo.test.harness;
 
 import org.apache.accumulo.core.client.security.tokens.AuthenticationToken;
-import org.apache.accumulo.harness.AccumuloClusterHarness.ClusterType;
+import org.apache.accumulo.test.harness.AccumuloClusterHarness.ClusterType;
 
 /**
  * Base functionality that must be provided as configuration to the test
